StreamingLayer-Spark-Kafka/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1

# Database Drivers
pymongo==4.6.1
dnspython==2.6.1

# Data Processing & Storage (Quan trọng cho GCS/Parquet)
pandas==2.2.0
pyarrow==15.0.0
gcsfs
fsspec